Adjusted Hammers to be a cube
Added Base Class for Hammer And Axe

Reworked Axe logic to be on use and utilise Item Orb
