+++ src/main/resources/assets/fabricated-exchange/lang/en_us.json
@@ -14,6 +14,7 @@
     "item.fabricated-exchange.medium_covalence_dust": "Medium Covalence Dust",
     "item.fabricated-exchange.high_covalence_dust": "High Covalence Dust",
     "item.fabricated-exchange.iron_band": "Iron Band",
+    "item.fabricated-exchange.item_orb": "",
     "item.fabricated-exchange.swiftwolfs_rending_gale": "Swiftwolf's Rending Gale",
     "item.fabricated-exchange.tome_of_knowledge": "Tome of Knowledge",
     "item.fabricated-exchange.dark_matter_sword": "Dark Matter Sword",
@@ -26,6 +27,9 @@
     "item.fabricated-exchange.dark_matter_hammer.mode_1": "3x3 Area",
     "item.fabricated-exchange.dark_matter_hammer.mode_2": "5x5 Area",
     "item.fabricated-exchange.dark_matter_axe": "Dark Matter Axe",
+    "item.fabricated-exchange.dark_matter_axe.mode_1": "Single Block",
+    "item.fabricated-exchange.dark_matter_axe.mode_2": "Tree Capitator",
+    "item.fabricated-exchange.dark_matter_axe.mode_3": "Strip",
     "item.fabricated-exchange.dark_matter_shovel": "Dark Matter Shovel",
     "item.fabricated-exchange.dark_matter_hoe": "Dark Matter Hoe",
     "item.fabricated-exchange.red_matter_sword": "Red Matter Sword",
@@ -47,8 +51,9 @@
     "item.fabricated-exchange.red_matter_hammer.mode_4": "9x9 Area",
     "item.fabricated-exchange.red_matter_axe": "Red Matter Axe",
     "item.fabricated-exchange.red_matter_axe.uncharged": "Disengaged",
-    "item.fabricated-exchange.red_matter_axe.mode_1": "Tree Capitator",
-    "item.fabricated-exchange.red_matter_axe.mode_2": "Single Block",
+    "item.fabricated-exchange.red_matter_axe.mode_1": "Single Block",
+    "item.fabricated-exchange.red_matter_axe.mode_2": "Tree Capitator",
+    "item.fabricated-exchange.red_matter_axe.mode_3": "Strip",
     "item.fabricated-exchange.red_matter_shovel": "Red Matter Shovel",
     "item.fabricated-exchange.red_matter_shovel.uncharged": "Inoperative",
     "item.fabricated-exchange.red_matter_shovel.mode_1": "Diggy Diggy Hole",
